update blog repository

+++ src/blogs/repositories/blogs.repository.ts
@@ -2,6 +2,9 @@ import {Blog} from "../types/blog";
 import {UpdateBlogDto} from "../dto/updateBlogDto";
 import {blogCollection} from "../../db/mongo.db";
 import { ObjectId, WithId } from 'mongodb';
+import {
+  RepositoryNotFoundError
+} from "../../core/errors/repositiry-not-found.error";
 
 export const blogsRepository = {
   // Найти все блоги
@@ -17,17 +20,27 @@ export const blogsRepository = {
     return await blogCollection.findOne({_id: new ObjectId(id)});
   },
 
+  // async findByIdOrFail(id: string): Promise<WithId<Blog>> {
+  //   const result = await blogCollection.findOne({_id: new ObjectId(id)});
+  //   if (!result) {
+  //     throw new RepositoryNotFoundError('No blog found with id ' + id);
+  //   }
+  //   return result
+  // },
+
   // Создать новый блог
-  async create(blog: Blog): Promise<WithId<Blog>| null> {
+  async create(blog: Blog): Promise<string> {
     const insertResult = await blogCollection.insertOne(blog);
 
-    return await blogCollection.findOne({
-      _id: insertResult.insertedId,
-    })
+    return insertResult.insertedId.toString();
   },
 
   // Обновить данные бдога
   async update(id: string, dto: UpdateBlogDto): Promise<boolean> {
+    if (!ObjectId.isValid(id)) {
+      return false
+    }
+
     const updateResult = await blogCollection.updateOne(
       {
         _id: new ObjectId(id),
@@ -45,12 +58,13 @@ export const blogsRepository = {
   },
 
   // Удалить блог
-  async delete(id: string): Promise<void> {
-    const deletedResult = await blogCollection.deleteOne({_id: new ObjectId(id)});
-    if (deletedResult.deletedCount < 1) {
-      throw new Error("Blog not exist");
+  async delete(id: string): Promise<boolean> {
+    if (!ObjectId.isValid(id)) {
+      return false
     }
-    return
-  },
 
+    const deletedResult = await blogCollection.deleteOne({_id: new ObjectId(id)});
+
+    return deletedResult.deletedCount === 1
+  },
 };
